Move TimelineCommand to server as it isn't meant to be shared
