+++ packages/reactotron-core-ui/src/reactotron-core-ui.ts
@@ -1,7 +1,6 @@
 import { Text } from "./text"
-import { TimelineCommand } from "./timeline-command"
 
 // Interfaces
 import { TimelineCommandOptions, ReactotronApp } from './reactotron-app';
 
-export { Text, TimelineCommand, TimelineCommandOptions, ReactotronApp }
+export { Text, TimelineCommandOptions, ReactotronApp }
