+++ packages/reactotron-server/src/app/screens/Timeline/timeline-command/timeline-command.tsx
@@ -1,6 +1,6 @@
 import React from "react"
+import { TimelineCommandOptions } from "reactotron-core-ui"
 
-import { TimelineCommandOptions } from "../reactotron-app"
 import { Timestamp } from "./timestamp"
 
 interface Props {
@@ -34,7 +34,7 @@ export class TimelineCommand extends React.Component<Props, State> {
     // with tailwind and flexbox, but I have no idea what it is.
     // So for the minute, please embrace the hack. Sorry!
     const style = {
-      minHeight: '10%'
+      minHeight: "10%"
     };
 
     return (
